Rename to gqcpy
