+++ gqcpy/setup.py.in
